requirements.txt
pyaudio
wave
openai
pyyaml